requirements.txt
fastapi
uvicorn
gitpython
mistralai
python-multipart
aiofiles
